Updated file structure to match V4 recommended
