+++ yo-cordapp/contracts-kotlin/build.gradle
@@ -1,0 +1,62 @@
+
+apply plugin: 'kotlin'
+apply plugin: 'net.corda.plugins.cordapp'
+apply plugin: 'net.corda.plugins.cordformation'
+apply plugin: 'net.corda.plugins.quasar-utils'
+
+cordapp {
+    targetPlatformVersion corda_platform_version.toInteger()
+    minimumPlatformVersion corda_platform_version.toInteger()
+    contract {
+        name "Yo CorDapp"
+        vendor "Corda Open Source"
+        licence "Apache License, Version 2.0"
+        versionId 1
+    }
+    workflow {
+        name "Yo CorDapp"
+        vendor "Corda Open Source"
+        licence "Apache License, Version 2.0"
+        versionId 1
+    }
+}
+
+
+sourceSets {
+    main {
+        resources {
+            srcDir "../config/dev"
+        }
+    }
+}
+
+    dependencies {
+        compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+        testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
+        testCompile "junit:junit:$junit_version"
+
+        // Corda dependencies.
+        cordaCompile "$corda_release_group:corda-core:$corda_release_version"
+        cordaCompile "$corda_release_group:corda-jackson:$corda_release_version"
+        cordaCompile "$corda_release_group:corda-rpc:$corda_release_version"
+        cordaRuntime "$corda_release_group:corda:$corda_release_version"
+
+        testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
+        cordaRuntime "org.apache.logging.log4j:log4j-slf4j-impl:$log4j_version"
+
+        // CorDapp dependencies.
+    }
+
+    jar {
+        // CorDapps do not configure a Node's logging.
+        exclude '**/log4j2*.xml'
+    }
+
+tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile) {
+    kotlinOptions {
+        languageVersion = "1.2"
+        apiVersion = "1.2"
+        jvmTarget = "1.8"
+        javaParameters = true   // Useful for reflection.
+    }
+}
