+++ yo-cordapp/contracts-kotlin/src/main/kotlin/net/corda/yo/contract/YoContract.kt
@@ -1,0 +1,29 @@
+package net.corda.yo.contract
+
+import net.corda.core.contracts.Contract
+import net.corda.core.contracts.TypeOnlyCommandData
+import net.corda.core.contracts.requireSingleCommand
+import net.corda.core.contracts.requireThat
+import net.corda.core.transactions.LedgerTransaction
+import net.corda.yo.state.YoState
+
+// Contract and state.
+class YoContract: Contract {
+    companion object {
+        // Used to identify our contract when building a transaction.
+        const val ID = "net.corda.yo.contract.YoContract"
+    }
+
+    // Command.
+    class Send : TypeOnlyCommandData()
+
+    // Contract code.
+    override fun verify(tx: LedgerTransaction) = requireThat {
+        val command = tx.commands.requireSingleCommand<Send>()
+        "There can be no inputs when Yo'ing other parties." using (tx.inputs.isEmpty())
+        "There must be one output: The Yo!" using (tx.outputs.size == 1)
+        val yo = tx.outputsOfType<YoState>().single()
+        "No sending Yo's to yourself!" using (yo.target != yo.origin)
+        "The Yo! must be signed by the sender." using (yo.origin.owningKey == command.signers.single())
+    }
+}
