+++ yo-cordapp/contracts-kotlin/src/test/kotlin/net/corda/yo/Main.kt
