+++ yo-cordapp/java-source/build.gradle
@@ -60,24 +60,14 @@ tasks.withType(JavaCompile) {
 }
 
 task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
-    nodeDefaults {
-        cordapp("$corda_release_group:corda-finance-contracts:$corda_release_version")
-        cordapp("$corda_release_group:corda-finance-workflows:$corda_release_version")
-        cordapp("$corda_release_group:corda-confidential-identities:$corda_release_version")
-        rpcUsers = [[ user: "user1", "password": "test", "permissions": ["ALL"]]]
-    }
     node {
         name "O=Notary,L=London,C=GB"
-        notary = [validating : false]
+        notary = [validating : true]
         p2pPort 10002
         rpcSettings {
             address("localhost:10003")
             adminAddress("localhost:10043")
         }
-        projectCordapp {
-            deploy = false
-        }
-        cordapps.clear()
     }
     node {
         name "O=PartyA,L=London,C=GB"
@@ -86,7 +76,7 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
             address("localhost:10006")
             adminAddress("localhost:10046")
         }
-        webPort 10007
+        rpcUsers = [[ user: "user1", "password": "test", "permissions": ["ALL"]]]
     }
     node {
         name "O=PartyB,L=New York,C=US"
@@ -95,15 +85,6 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
             address("localhost:10009")
             adminAddress("localhost:10049")
         }
-        webPort 10010
-    }
-    node {
-        name "O=PartyC,L=Paris,C=FR"
-        p2pPort 10011
-        rpcSettings {
-            address("localhost:10012")
-            adminAddress("localhost:10052")
-        }
-        webPort 10013
+        rpcUsers = [[ user: "user1", "password": "test", "permissions": ["ALL"]]]
     }
 }
