--- yo-cordapp/kotlin-source/build.gradle
@@ -1,123 +1,0 @@
-buildscript {
-    Properties constants = new Properties()
-    file("$projectDir/../../constants.properties").withInputStream { constants.load(it) }
-
-    ext {
-        corda_release_group = constants.getProperty("cordaReleaseGroup")
-        corda_release_version = constants.getProperty("cordaVersion")
-        corda_gradle_plugins_version = constants.getProperty("gradlePluginsVersion")
-        kotlin_version = constants.getProperty("kotlinVersion")
-        junit_version = constants.getProperty("junitVersion")
-        quasar_version = constants.getProperty("quasarVersion")
-        log4j_version = constants.getProperty("log4jVersion")
-        slf4j_version = constants.getProperty("slf4jVersion")
-        corda_platform_version = constants.getProperty("platformVersion")
-    }
-
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        jcenter()
-        maven { url 'https://ci-artifactory.corda.r3cev.com/artifactory/corda-releases' }
-    }
-
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath "net.corda.plugins:cordapp:$corda_gradle_plugins_version"
-        classpath "net.corda.plugins:cordformation:$corda_gradle_plugins_version"
-        classpath "net.corda.plugins:quasar-utils:$corda_gradle_plugins_version"
-    }
-}
-
-apply from: '../repositories.gradle'
-
-apply plugin: 'kotlin'
-apply plugin: 'net.corda.plugins.cordapp'
-apply plugin: 'net.corda.plugins.cordformation'
-apply plugin: 'net.corda.plugins.quasar-utils'
-
-cordapp {
-    targetPlatformVersion corda_platform_version.toInteger()
-    minimumPlatformVersion corda_platform_version.toInteger()
-    contract {
-        name "Yo CorDapp"
-        vendor "Corda Open Source"
-        licence "Apache License, Version 2.0"
-        versionId 1
-    }
-    workflow {
-        name "Yo CorDapp"
-        vendor "Corda Open Source"
-        licence "Apache License, Version 2.0"
-        versionId 1
-    }
-}
-
-sourceSets {
-    main {
-        resources {
-            srcDir "config/dev"
-        }
-    }
-
-    dependencies {
-        compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-        testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
-        testCompile "junit:junit:$junit_version"
-
-        // Corda dependencies.
-        cordaCompile "$corda_release_group:corda-core:$corda_release_version"
-        cordaCompile "$corda_release_group:corda-jackson:$corda_release_version"
-        cordaCompile "$corda_release_group:corda-rpc:$corda_release_version"
-        cordaRuntime "$corda_release_group:corda:$corda_release_version"
-
-        testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
-        cordaRuntime "org.apache.logging.log4j:log4j-slf4j-impl:$log4j_version"
-
-        // CorDapp dependencies.
-    }
-
-    jar {
-        // CorDapps do not configure a Node's logging.
-        exclude '**/log4j2*.xml'
-    }
-
-    tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile) {
-        kotlinOptions {
-            languageVersion = "1.2"
-            apiVersion = "1.2"
-            jvmTarget = "1.8"
-            javaParameters = true   // Useful for reflection.
-        }
-    }
-
-    task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
-        node {
-            name "O=Notary,L=London,C=GB"
-            notary = [validating: true]
-            p2pPort 10002
-            rpcSettings {
-                address("localhost:10003")
-                adminAddress("localhost:10043")
-            }
-        }
-        node {
-            name "O=PartyA,L=London,C=GB"
-            p2pPort 10005
-            rpcSettings {
-                address("localhost:10006")
-                adminAddress("localhost:10046")
-            }
-            rpcUsers = [[user: "user1", "password": "test", "permissions": ["ALL"]]]
-        }
-        node {
-            name "O=PartyB,L=New York,C=US"
-            p2pPort 10008
-            rpcSettings {
-                address("localhost:10009")
-                adminAddress("localhost:10049")
-            }
-            rpcUsers = [[user: "user1", "password": "test", "permissions": ["ALL"]]]
-        }
-    }
-}
