+++ yo-cordapp/settings.gradle
@@ -1,3 +1,4 @@
 
-include 'kotlin-source'
-include 'java-source'
+include 'contracts-kotlin'
+//include 'java-source'
+include 'workflows-kotlin'
