+++ yo-cordapp/workflows-kotlin/src/main/kotlin/net.corda.yo.flow/YoFlow.kt
@@ -1,13 +1,15 @@
-package net.corda.yo
+package net.corda.yo.flow
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.core.contracts.*
+import net.corda.core.contracts.Command
+import net.corda.core.contracts.StateAndContract
 import net.corda.core.flows.*
 import net.corda.core.identity.Party
-import net.corda.core.transactions.LedgerTransaction
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.TransactionBuilder
 import net.corda.core.utilities.ProgressTracker
+import net.corda.yo.contract.YoContract
+import net.corda.yo.state.YoState
 
 // Flow.
 @InitiatingFlow
@@ -56,34 +58,4 @@ class YoFlowResponder(val counterpartySession: FlowSession) : FlowLogic<SignedTr
     override fun call(): SignedTransaction {
         return subFlow(ReceiveFinalityFlow(counterpartySession))
     }
-}
-
-// Contract and state.
-class YoContract: Contract {
-    companion object {
-        // Used to identify our contract when building a transaction.
-        const val ID = "net.corda.yo.YoContract"
-    }
-
-    // Command.
-    class Send : TypeOnlyCommandData()
-
-    // Contract code.
-    override fun verify(tx: LedgerTransaction) = requireThat {
-        val command = tx.commands.requireSingleCommand<Send>()
-        "There can be no inputs when Yo'ing other parties." using (tx.inputs.isEmpty())
-        "There must be one output: The Yo!" using (tx.outputs.size == 1)
-        val yo = tx.outputsOfType<YoState>().single()
-        "No sending Yo's to yourself!" using (yo.target != yo.origin)
-        "The Yo! must be signed by the sender." using (yo.origin.owningKey == command.signers.single())
-    }
-}
-
-// State.
-@BelongsToContract(YoContract::class)
-data class YoState(val origin: Party,
-                 val target: Party,
-                 val yo: String = "Yo!") : ContractState {
-    override val participants = listOf(target)
-    override fun toString() = "${origin.name}: $yo"
 }
