+++ yo-cordapp/workflows-kotlin/src/test/kotlin/net.corda.yo.test/flow/YoFlowTests.kt
@@ -1,0 +1,55 @@
+package net.corda.yo.test.flow
+
+import net.corda.core.node.services.queryBy
+import net.corda.core.utilities.getOrThrow
+import net.corda.testing.node.MockNetwork
+import net.corda.testing.node.MockNetworkParameters
+import net.corda.testing.node.StartedMockNode
+import net.corda.testing.node.TestCordapp
+import net.corda.yo.flow.YoFlow
+import net.corda.yo.state.YoState
+import org.junit.After
+import org.junit.Before
+import org.junit.Test
+import kotlin.test.assertEquals
+
+class YoFlowTests {
+    private lateinit var network: MockNetwork
+    private lateinit var a: StartedMockNode
+    private lateinit var b: StartedMockNode
+
+    @Before
+    fun setup() {
+        network = MockNetwork(MockNetworkParameters(cordappsForAllNodes = listOf(
+                TestCordapp.findCordapp("net.corda.yo.flow"),
+                TestCordapp.findCordapp("net.corda.yo.contract"))))
+        a = network.createPartyNode()
+        b = network.createPartyNode()
+        network.runNetwork()
+    }
+
+    @After
+    fun tearDown() {
+        network.stopNodes()
+    }
+
+    @Test
+    fun flowWorksCorrectly() {
+        val yo = YoState(a.info.legalIdentities.first(), b.info.legalIdentities.first())
+        val flow = YoFlow(b.info.legalIdentities.first())
+        val future = a.startFlow(flow)
+        network.runNetwork()
+        val stx = future.getOrThrow()
+        // Check yo transaction is stored in the storage service.
+        val bTx = b.services.validatedTransactions.getTransaction(stx.id)
+        assertEquals(bTx, stx)
+        println("bTx == $stx")
+        // Check yo state is stored in the vault.
+        b.transaction {
+            // Simple query.
+            val bYo = b.services.vaultService.queryBy<YoState>().states.single().state.data
+            assertEquals(bYo.toString(), yo.toString())
+            println("$bYo == $yo")
+        }
+    }
+}
